refactor: move non-root components into a dir
