+++ packages/react-tinacms-inline/src/inline-field-image.tsx
@@ -19,8 +19,8 @@ limitations under the License.
 import * as React from 'react'
 import { InlineField } from './inline-field'
 import { useCMS, Form } from 'tinacms'
-import { useDropzone } from 'react-dropzone'
 import { InputFocusWrapper } from './styles'
+import { ImageUpload } from './components/image-upload'
 
 export interface InlineImageProps {
   name: string
@@ -70,36 +70,5 @@ export function InlineImageField({
         return children ? children : <img src={input.value} />
       }}
     </InlineField>
-  )
-}
-
-export interface ImageUploadProps {
-  onDrop: (acceptedFiles: any[]) => void
-  value?: string
-  children?: any
-}
-
-export const ImageUpload = ({ onDrop, value, children }: ImageUploadProps) => {
-  const {
-    getRootProps,
-    getInputProps,
-    isDragActive,
-    isDragAccept,
-    isDragReject,
-  } = useDropzone({ accept: 'image/*', onDrop })
-
-  return (
-    <div {...getRootProps({ isDragActive, isDragAccept, isDragReject })}>
-      <input {...getInputProps()} />
-      {value ? (
-        <div>{children ? children : <img src={value} />}</div>
-      ) : (
-        <div>
-          Drag 'n' drop some files here,
-          <br />
-          or click to select files
-        </div>
-      )}
-    </div>
   )
 }
